+++ examples/simple_pipelayer/src/app/filter/hello_filter.py
@@ -1,7 +1,9 @@
+from typing import Any
+
 from app.app_context import AppContext
 from pipelayer import Filter
 
 
 class HelloFilter(Filter):
-    def run(self, context: AppContext, data: str = None) -> str:
+    def run(self, data: Any, context: AppContext) -> str:
         return "Hello"