+++ src/pipelayer/filter.py
@@ -4,14 +4,22 @@ from abc import abstractmethod
 from typing import Any, Callable, Optional
 
 from pipelayer.context import Context
-from pipelayer.step import Step
 
 
-class Filter(Step):
-    def __init__(self: Filter, name: str = "", pre_process: Callable = None, post_process: Callable = None) -> None:
-        super().__init__(name)
+class Filter:
+    def __init__(
+        self: Filter,
+        name: str = "",
+        pre_process: Optional[Callable[[Any, Context], Any]] = None,
+        post_process: Optional[Callable[[Any, Context], Any]] = None
+    ) -> None:
+        self.__name = name or self.__class__.__name__
         self.__pre_process = pre_process
         self.__post_process = post_process
+
+    @property
+    def name(self) -> str:
+        return self.__name
 
     @property
     def pre_process(self) -> Optional[Callable]:
@@ -22,5 +30,5 @@ class Filter(Step):
         return self.__post_process
 
     @abstractmethod
-    def run(self, data: Any, context: Context) -> Any:
+    def run(self, data: Any, context: Any) -> Any:
         raise NotImplementedError